Annotation permission model

+++ annotationengine/api.py
@@ -9,19 +9,24 @@ from dynamicannotationdb.errors import (
 from flask import Response, abort, g, request
 from flask_accepts import accepts
 from flask_restx import Namespace, Resource, reqparse
-from middle_auth_client import auth_requires_admin, auth_requires_permission
+from middle_auth_client import (
+    auth_requires_admin,
+    auth_requires_permission,
+    users_share_common_group,
+)
 
 from annotationengine.aligned_volume import get_aligned_volumes
-from annotationengine.anno_database import get_db
+from annotationengine.anno_database import get_db, check_write_permission, check_read_permission, check_ownership
 from annotationengine.errors import SchemaServiceError
 from annotationengine.schemas import (
     CreateTableSchema,
     DeleteAnnotationSchema,
     FullMetadataSchema,
     PutAnnotationSchema,
+    UpdateMetadataSchema,
 )
 
-from .api_examples import synapse_table_example
+from .api_examples import synapse_table_example, synapse_table_update_example
 
 __version__ = "4.1.1"
 
@@ -100,6 +105,24 @@ class Table(Resource):
         return Response(table_info, headers=headers)
 
     @auth_requires_permission(
+        "edit", table_arg="aligned_volume_name", resource_namespace="aligned_volume"
+    )
+    @api_bp.doc("update_table", security="apikey", example=synapse_table_update_example)
+    @accepts("UpdateMetadataSchema", schema=UpdateMetadataSchema, api=api_bp)
+    def put(self, aligned_volume_name: str)-> FullMetadataSchema:
+        data = request.parsed_obj
+        db = get_db(aligned_volume_name)
+        metadata_dict = data.get("metadata")
+        if metadata_dict.get("user_id", None) is None:
+            metadata_dict["user_id"] = str(g.auth_user["id"])
+
+        table_name = data.get("table_name")
+        check_ownership(db, table_name)
+        
+        new_md= db.annotation.update_table_metadata(table_name, **metadata_dict)
+        return new_md, 200
+
+    @auth_requires_permission(
         "view", table_arg="aligned_volume_name", resource_namespace="aligned_volume"
     )
     @api_bp.doc("get_aligned_volume_tables", security="apikey")
@@ -123,17 +146,24 @@ class AnnotationTable(Resource):
         """Get metadata for a given table"""
         check_aligned_volume(aligned_volume_name)
         db = get_db(aligned_volume_name)
-        return db.database.get_table_metadata(table_name), 200
+        md = db.database.get_table_metadata(table_name)
+        headers = None
+        if md.get('notice_text',None) is not None:
+            headers={"Warning":f"Table Owner Notice: {md['notice_text']}"}
+        return md, 201, headers
 
-    @auth_requires_admin
+    @auth_requires_permission(
+        "edit", table_arg="aligned_volume_name", resource_namespace="aligned_volume"
+    )
     @api_bp.doc(
-        description="mark an annotation table for deletion (admin only)",
+        description="mark an annotation table for deletion",
         security="apikey",
     )
     def delete(self, aligned_volume_name: str, table_name: str) -> bool:
         """Delete an annotation table (marks for deletion, will suspend materialization, admin only)"""
         check_aligned_volume(aligned_volume_name)
         db = get_db(aligned_volume_name)
+        check_ownership(db, table_name)
         is_deleted = db.annotation.delete_table(table_name)
         return is_deleted, 200
 
@@ -150,6 +180,7 @@ class TableInfo(Resource):
         """Get count of rows of an annotation table"""
         check_aligned_volume(aligned_volume_name)
         db = get_db(aligned_volume_name)
+        check_read_permission(db, table_name)
         return db.database.get_annotation_table_size(table_name), 200
 
 
@@ -165,11 +196,16 @@ class Annotations(Resource):
     def get(self, aligned_volume_name: str, table_name: str, **kwargs):
         """Get annotations by list of IDs"""
         check_aligned_volume(aligned_volume_name)
+        db = get_db(aligned_volume_name)
+        check_read_permission(db, table_name)
         args = annotation_parser.parse_args()
 
         annotation_ids = args["annotation_ids"]
 
-        db = get_db(aligned_volume_name)
+        md = db.database.get_table_metadata(table_name)
+        headers = None
+        if md.get('notice_text',None) is not None:
+            headers={"Warning":f"Table Owner Notice: {md['notice_text']}"}
 
         annotations = db.annotation.get_annotations(table_name, annotation_ids)
 
@@ -177,7 +213,7 @@ class Annotations(Resource):
             msg = f"annotation_id {annotation_ids} not in {table_name}"
             abort(404, msg)
 
-        return annotations, 200
+        return annotations, 200, headers
 
     @auth_requires_permission(
         "edit", table_arg="aligned_volume_name", resource_namespace="aligned_volume"
@@ -188,9 +224,7 @@ class Annotations(Resource):
         """Insert annotations"""
         check_aligned_volume(aligned_volume_name)
         db = get_db(aligned_volume_name)
-        metadata = db.database.get_table_metadata(table_name)
-        if metadata["user_id"] != str(g.auth_user["id"]):
-            return Response("Unauthorized: You did not create this table", 401)
+        check_write_permission(db, table_name)
         data = request.parsed_obj
         annotations = data.get("annotations")
 
@@ -214,9 +248,7 @@ class Annotations(Resource):
         """Update annotations"""
         check_aligned_volume(aligned_volume_name)
         db = get_db(aligned_volume_name)
-        metadata = db.database.get_table_metadata(table_name)
-        if metadata["user_id"] != str(g.auth_user["id"]):
-            return Response("Unauthorized: You did not create this table", 401)
+        check_write_permission(db, table_name)
         data = request.parsed_obj
 
         annotations = data.get("annotations")
@@ -243,15 +275,11 @@ class Annotations(Resource):
         """Delete annotations"""
         check_aligned_volume(aligned_volume_name)
         db = get_db(aligned_volume_name)
-        metadata = db.database.get_table_metadata(table_name)
-        if metadata["user_id"] != str(g.auth_user["id"]):
-            return Response("Unauthorized: You did not create this table", 401)
+        check_write_permission(db, table_name)
+       
         data = request.parsed_obj
 
         ids = data.get("annotation_ids")
-
-        db = get_db(aligned_volume_name)
-
         deleted_ids = db.annotation.delete_annotation(table_name, ids)
 
         if deleted_ids is None:
